@@ -2,6 +2,7 @@
   "name": "press-kit-ad",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "next dev",
     "build": "NODE_ENV=production next build",
@@ -12,22 +13,23 @@
   "dependencies": {
     "@tailwindcss/cli": "^4.1.13",
     "@tailwindcss/postcss": "^4.1.13",
-    "@types/node": "20.14.9",
-    "@types/react": "18.3.3",
-    "@types/react-dom": "18.3.0",
-    "autoprefixer": "10.4.19",
-    "eslint": "^8.6.0",
-    "eslint-config-next": "14.2.4",
-    "next": "14.2.4",
-    "postcss": "8.4.39",
-    "react": "18.3.1",
-    "react-dom": "18.3.1",
-    "react-player": "^2.16.0",
-    "react-scroll": "^1.9.0",
+    "@types/node": "^24.3.1",
+    "@types/react": "^19.1.12",
+    "@types/react-dom": "^19.1.9",
+    "autoprefixer": "^10.4.21",
+    "eslint": "^9.35.0",
+    "eslint-config-next": "^15.5.3",
+    "next": "^15.5.3",
+    "postcss": "^8.5.6",
+    "react": "^19.1.1",
+    "react-dom": "^19.1.1",
+    "react-player": "^3.3.2",
+    "react-scroll": "^1.9.3",
     "tailwindcss": "4.1.13",
-    "typescript": "5.5.3"
+    "typescript": "^5.9.2"
   },
   "devDependencies": {
+    "@eslint/eslintrc": "^3.2.0",
     "@types/react-scroll": "^1.8.10"
   }
 }
