@@ -1,24 +1,31 @@
 "use client";
-import { useEffect, useState } from "react";
-import ReactPlayer from "react-player/youtube";
+import React, { useEffect, useState } from "react";
+import ReactPlayer from "react-player";
 
-export default function Landing() {
+export default function LandingBg() {
   const [isLoaded, setIsLoaded] = useState(false);
 
   useEffect(() => {
     setIsLoaded(true);
   }, []);
+
   return (
-    <div className="fixed left-0 right-0 h-screen z-0 brightness-0.4 scale-150">
+    <div className="fixed left-0 right-0 h-screen z-0 brightness-0.4 scale-150 overflow-hidden">
       {isLoaded ? (
         <ReactPlayer
-          playbackRate={0.8}
+          width="100%"
+          height="100%"
+          src="https://youtu.be/oB325uTDKIw?si=FbP7X-cNp9hmUk2M"
           playing={true}
           loop={true}
+          playbackRate={0.8}
           muted={true}
-          width={"100%"}
-          height={"100%"}
-          url="https://youtu.be/oB325uTDKIw?si=FbP7X-cNp9hmUk2M"
+          controls={false}
+          className="absolute top-0 left-0"
+          style={{
+            objectFit: "cover",
+            transform: "scale(1.1)",
+          }}
         />
       ) : null}
     </div>
