@@ -1,4 +1,4 @@
-import { ReactElement } from "react";
+import type { ReactElement } from "react";
 
 export const Section = ({
   children,
@@ -14,11 +14,10 @@ export const Section = ({
   return (
     <section
       id={id}
-      className={` w-full flex  justify-center items-center relative z-2 ${
-        bg
+      className={` w-full flex  justify-center items-center relative z-2 ${bg
           ? "bg-gradient-to-b from-bubblegum/20 to-blue/20 backdrop-blur-md"
           : ""
-      }
+        }
        ${fullVh ? "h-screen" : ""}`}
     >
       {/* <div className="rounded-3xl bg-bubblegum/20 backdrop-blur-lg w-full lg:mx-24 xl:mx-48 p-4 md:p-8 lg:p-16 m-8 lg:my-16">
